+++ packages/create-sitecore-jss/src/templates/nextjs-personalize/src/components/CdpIntegrationScript.tsx
@@ -1,4 +1,4 @@
-import { useSitecoreContext, parsePointOfSaleRawInput } from '@sitecore-jss/sitecore-jss-nextjs';
+import { useSitecoreContext, parseMultiValuePointOfSale, shouldHandleMultiValuePos } from '@sitecore-jss/sitecore-jss-nextjs';
 import Script from 'next/script';
 import { useEffect } from 'react';
 import config from 'temp/config';
@@ -21,11 +21,14 @@ interface BoxeverViewEventArgs {
 }
 
 function createPageView(locale: string, routeName: string) {
+  try {
   // POS can be multi-valued (one entry per locale) or single valued so we parse it
   // POS must be valid in order to save events (domain name might be taken but it must be defined in CDP settings)
   // Later on point of sale should be available from layout service data - using env for now
-  const parsedPos = parsePointOfSaleRawInput(process.env.NEXT_PUBLIC_CDP_POINTOFSALE || window.location.host.replace(/^www\./, ''));
-  const pointOfSale = parsedPos[locale] ?? parsedPos['default'];
+  const parsedPos = (process.env.NEXT_PUBLIC_CDP_POINTOFSALE && shouldHandleMultiValuePos(process.env.NEXT_PUBLIC_CDP_POINTOFSALE)) ? 
+    parseMultiValuePointOfSale(process.env.NEXT_PUBLIC_CDP_POINTOFSALE)[locale] : 
+    process.env.NEXT_PUBLIC_CDP_POINTOFSALE
+  const pointOfSale = parsedPos || window.location.host.replace(/^www\./, '');
 
   _boxeverq.push(function () {
     const pageViewEvent: BoxeverViewEventArgs = {
@@ -45,6 +48,12 @@ function createPageView(locale: string, routeName: string) {
       'json'
     );
   });
+}
+catch (error) {
+  // if we have invalid data or request is failing we should abort event creation
+  console.log('createPageView failed');
+  console.log(error);
+}
 }
 
 const CdpIntegrationScript = (): JSX.Element => {
