+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -55,6 +55,8 @@ export {
   getPersonalizedRewrite,
   getPersonalizedRewriteData,
   normalizePersonalizedRewrite,
+  shouldHandleMultiValuePos,
+  parseMultiValuePointOfSale,
 } from '@sitecore-jss/sitecore-jss/personalize';
 export {
   RobotsQueryResult,
