+++ packages/sitecore-jss/src/personalize/utils.test.ts
@@ -3,6 +3,7 @@ import {
   getPersonalizedRewrite,
   getPersonalizedRewriteData,
   normalizePersonalizedRewrite,
+  parseMultiValuePointOfSale,
   VARIANT_PREFIX,
   DEFAULT_VARIANT,
 } from './utils';
@@ -76,6 +77,24 @@ describe('utils', () => {
       const pathname = `/${VARIANT_PREFIX}foo`;
       const result = normalizePersonalizedRewrite(pathname);
       expect(result).to.equal('/');
+    });
+  });
+  describe('parseMultiValuePointOfSale', () => {
+    it('should return Record when valid json collection provided', () => {
+      const cdpRecord: Record<string, string> = {
+        en: 'value1',
+        'da-DA': 'nke',
+      };
+      const validJson = JSON.stringify(cdpRecord);
+
+      const result = parseMultiValuePointOfSale(validJson);
+
+      expect(result).to.deep.equal(cdpRecord);
+    });
+    it('should throw when input value with invlaid json provided', () => {
+      const invalidJson = '{abcderef}';
+
+      expect(parseMultiValuePointOfSale.bind(null, invalidJson)).to.throw();
     });
   });
 });
